form validation
Summary: This is the module used for form validation. It's aim is to be simple to use, initiating it on the server, it will provide an endpoint for the client side script, along with endpoints for checking if an asynchronous validation function, and checking al fields when the form is submitted.

Test Plan: npm test

Reviewers: steve

CC: Engineering

Differential Revision: https://phabricator.poptip.com/D77